+++ atlas/static/mapAreas.js
@@ -65,7 +65,7 @@ htmlLegend = configuration.AFFICHAGE_MAILLE
 generateLegende(htmlLegend);
 
 function displayObsPreciseBaseUrl() {
-    if (sheetType === 'commune') {
+    if (window.location.pathname === `/zone/${areaInfos.areaCode}`) {
         return configuration.URL_APPLICATION + "/api/observations/" + areaInfos.areaCode
     } else {
         return configuration.URL_APPLICATION + "/api/observations/area/" + areaInfos.id_area
@@ -96,13 +96,13 @@ function displayObsPreciseBaseUrl(areaCode, cd_ref) {
         if (configuration.AFFICHAGE_MAILLE) {
             displayMailleLayerLastObs(observations);
         } else {
-            displayMarkerLayerPointCommune(observations);
+            displayMarkerLayerPointZone(observations);
         }
     });
 }
 
 function displayObsGridBaseUrl() {
-    if (sheetType === 'commune') {
+    if (window.location.pathname === `/zone/${areaInfos.areaCode}`) {
         return configuration.URL_APPLICATION + "/api/observationsMaille/"
     } else {
         return configuration.URL_APPLICATION + "/api/observationsMaille/area/"
@@ -110,12 +110,12 @@ function displayObsGridBaseUrl() {
 }
 
 // display observation on click
-function displayObsTaxon(insee, cd_ref) {
+function displayObsTaxon(id_zone, cd_ref) {
   $.ajax({
     url:
       configuration.URL_APPLICATION +
       "/api/observations/" +
-      insee +
+      id_zone +
       "/" +
       cd_ref,
     dataType: "json",
@@ -132,7 +132,7 @@ function displayObsTaxon(insee, cd_ref) {
     if (configuration.AFFICHAGE_MAILLE) {
       displayMailleLayerLastObs(observations);
     } else {
-      displayMarkerLayerPointCommune(observations);
+      displayMarkerLayerPointZone(observations);
     }
   });
 }
@@ -164,9 +164,9 @@ function refreshObsArea() {
             .removeClass("current");
         $(this).addClass("current");
         if (configuration.AFFICHAGE_MAILLE) {
-            displayObsTaxonMaille($(this).attr("area-code"), $(this).attr("cdRef"));
+            displayObsTaxonMaille(this.getAttribute("area-code"), this.getAttribute("cdref"));
         } else {
-            displayObsTaxon($(this).attr("area-code"), $(this).attr("cdRef"));
+            displayObsTaxon(this.getAttribute("area-code"), this.getAttribute("cdref"));
         }
         var name = $(this)
             .find("#name")