+++ data/atlas/12.atlas.t_mailles_territoire.sql
@@ -9,7 +9,7 @@ CREATE TABLE atlas.t_mailles_territoire AS
         JOIN ref_geo.bib_areas_types AS t
             ON t.id_type = a.id_type
         JOIN atlas.t_layer_territoire AS l
-            ON ST_intersects(a.geom, st_transform(l.the_geom, find_srid('ref_geo', 'l_areas', 'geom')))
+            ON ST_intersects(a.geom, st_transform(l.geom, find_srid('ref_geo', 'l_areas', 'geom')))
     WHERE a.enable = true
         AND t.type_code = :type_maille ;