+++ data/atlas/8.atlas.vm_medias.sql
@@ -12,7 +12,6 @@ CREATE MATERIALIZED VIEW atlas.vm_medias AS
       t_medias.id_type,
       t_medias.licence,
       t_medias.source
-   FROM taxonomie.t_medias
-WHERE NOT t_medias.supprime = true;
+   FROM taxonomie.t_medias;
 
 CREATE UNIQUE INDEX ON atlas.vm_medias(id_media);